Refactor panel hit result

+++ src/main/kotlin/com/xfastgames/witness/screens/solver/PuzzleSolverViewModels.kt
@@ -21,7 +21,7 @@ sealed class PuzzleSolverData {
 }
 
 data class PuzzlePanelHitResult(
-    val position: Pair<Double, Double>,
+    val position: Pair<Float, Float>,
     val puzzlePanel: Panel,
     val blockEntity: PuzzleFrameBlockEntity,
 )
